Update tts_engine field to use SmallIntegerField

+++ ayushma/migrations/0051_project_tts_engine.py
@@ -12,7 +12,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name="project",
             name="tts_engine",
-            field=models.IntegerField(
+            field=models.SmallIntegerField(
                 choices=[(1, "Openai"), (2, "Google")], default=2
             ),
         ),
